Dynamic kernel: optensor (#349)
* dynamic kernels optensor: fwdbias/generic
* dynamic kernels optensor: leadingones/generic
* dynamic kernels optensor: squash2d
* dynamic kernels optensor: 2d/4d lite
* update optensor network config

[ROCm/MIOpen commit: 05f0dfe2c7994b47fe3cc7aa4ca78b7d9637d367]

+++ src/ocl/rnn_util_ocl.cpp
@@ -68,7 +68,7 @@ void LSTMForwardHiddenStateUpdate(const Handle& handle,
                          ? 4
                          : ((total_work >= 2 * max_active_threads && hy_h % 2 == 0) ? 2 : 1);
 
-    size_t total_item   = std::max(size_t(total_work / RD_BLCK), size_t(1));
+    size_t total_item   = std::max(total_work / RD_BLCK, size_t(1));
     size_t item_per_grp = total_item <= 64 ? 64 : total_item <= 128 ? 128 : 256;
     size_t glb_sz       = total_item < max_active_threads ? total_item : max_active_threads;
     size_t wg_sz        = (glb_sz + item_per_grp - 1) / item_per_grp;
@@ -196,7 +196,7 @@ void LSTMBackwardHiddenStateUpdate(const Handle& handle,
                          ? 4
                          : ((total_work >= 2 * max_active_threads && hy_h % 2 == 0) ? 2 : 1);
 
-    size_t total_item   = std::max(size_t(total_work / RD_BLCK), size_t(1));
+    size_t total_item   = std::max(total_work / RD_BLCK, size_t(1));
     size_t item_per_grp = total_item <= 64 ? 64 : total_item <= 128 ? 128 : 256;
     size_t glb_sz       = total_item < max_active_threads ? total_item : max_active_threads;
     size_t wg_sz        = (glb_sz + item_per_grp - 1) / item_per_grp;
